moved profiling to utils
