+++ raiden/benchmark/profile_transfer.py
@@ -5,9 +5,8 @@ from collections import namedtuple
 
 from ethereum import slogging
 
-from raiden import profiling
 from raiden.app import DEFAULT_SETTLE_TIMEOUT
-from raiden.utils import sha3
+from raiden.utils import sha3, profiling
 from raiden.tests.utils.network import create_network
 from raiden.network.transport import UDPTransport
 from raiden.tests.utils.mock_client import (
