+++ raiden/utils/profiling/stack.py
